refactor: use npm instead of yarn
Recently, Yarn 2.0 is released and it is now a bit more complex to install.
It is still OK in my personal projects. Even in the team projects, it should not be a big problem for JSers and frontend engineers. However, non-JS experts may be confused by it.

Yarn 2.0 also adds files which needs to be committed to the Git repository. I don't like it.

Therefore, I decided to use npm for new projects.